# openApi.yaml
openapi: 3.0.3
info:
  title: E-Commerce API
  version: 1.0.0
  description: A RESTful API Built with Node/Express and PostgreSQL for an E-Commerce App.
  contact:
    email: nimazarei53@gmail.com
    name: the Developer
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
...
servers:
  - url: https://e-commerce-pern.herokuapp.com
paths:
...
  /api/v1/carts:
    get:
      summary: Get All of the Carts' Data (ADMIN)
      description: Retrieve all of the users' carts information from the database.
      operationId: getCarts
      responses:
        "200":
          description: Successfully Retrieved Carts' Data
        "401":
          description: Authorization Error
        "500":
          description: Internal Server Error
      tags:
        - Carts
  /api/v1/carts/{cart_id}:
    post:
      summary: Add Product to Cart (OWNER)
      description: Add a product to a user's cart matching the provided cart ID.
      operationId: addProductToCart
      parameters:
        - $ref: "#/components/parameters/cartParam"
      requestBody:
        description: Product Values
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Cart"
            example:
              product_id: bd38abae-420a-4fbf-97f1-fbb316726e1f
              quantity: 1
              color: Aura Black
      responses:
        "201":
          description: Successfully Added Product to Cart
          content:
            application/json:
              examples:
                cart:
                  value: |-
                    {
                        "message": "Product Added to Cart"
                    }
        "400":
          description: Validation Error
        "401":
          description: Authorization Error
        "404":
          description: Not Found Error
        "500":
          description: Internal Server Error
      tags:
        - Carts
    put:
      summary: Update a Product's Quantity in Cart (OWNER)
      description: >-
        Update a product's quantity in the user's cart matching the provided
        cart ID.
      operationId: updateCartProduct
      parameters:
        - $ref: "#/components/parameters/cartParam"
      requestBody:
        description: Update Values
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Cart"
            example:
              product_id: bd38abae-420a-4fbf-97f1-fbb316726e1f
              quantity: 3
      responses:
        "200":
          description: Successfully Updated Product's Quantity in Cart
          content:
            application/json:
              examples:
                cart:
                  value: |-
                    {
                        "message": "Product Quantity Updated Successfully"
                    }
        "400":
          description: Validation Error
        "401":
          description: Authorization Error
        "404":
          description: Not Found Error
        "500":
          description: Internal Server Error
      tags:
        - Carts
  /api/v1/carts/{user_id}:
    get:
      summary: Get a User's Products from Cart (ADMIN / OWNER)
      description: Retrieve the products from a user's cart matching the provided user ID.
      operationId: getCartProducts
      parameters:
        - $ref: "#/components/parameters/userParam"
      responses:
        "200":
          description: Successfully Retrieved Products from Cart
          content:
            application/json:
              examples:
                cart:
                  value: |-
                    [
                        {
                            "product_id": "81294da5-f1fe-4d5d-8da2-27c8a4dd0a38",
                            "name": "Samsung Galaxy Note 8",
                            "price": "219.99",
                            "img_urls": [
                                "https://res.cloudinary.com/de3kst1ud/image/upload/v1650202227/Galaxy%20Note%208/Samsung_Galaxy_Note_8_-_1_hbaavf.jpg",
                                "https://res.cloudinary.com/de3kst1ud/image/upload/v1650202229/Galaxy%20Note%208/Samsung_Galaxy_Note_8_-_2_vr91ku.jpg",
                                "https://res.cloudinary.com/de3kst1ud/image/upload/v1650202227/Galaxy%20Note%208/Samsung_Galaxy_Note_8_-_3_euojga.jpg"
                            ],
                            "quantity": 1,
                            "color": "Midnight Black"
                        }
                    ]
        "401":
          description: Authorization Error
        "404":
          description: Not Found Error
        "500":
          description: Internal Server Error
      tags:
        - Carts
  /api/v1/carts/{cart_id}/products/{product_id}:
    delete:
      summary: Delete a Product from Cart (OWNER)
      description: >-
        Delete a product from the user's cart matching the provided cart and
        product IDs.
      operationId: deleteCartProduct
      parameters:
        - $ref: "#/components/parameters/cartParam"
        - $ref: "#/components/parameters/productParam"
      responses:
        "200":
          description: Successfully Deleted Product from Cart
          content:
            application/json:
              examples:
                cart:
                  value: |-
                    {
                        "message": "Product Deleted from Cart"
                    }
        "401":
          description: Authorization Error
        "404":
          description: Not Found Error
        "500":
          description: Internal Server Error
      tags:
        - Carts
  /api/v1/carts/{cart_id}/checkout:
    post:
      summary: Check Out Cart (OWNER)
      description: Check out the products in cart, create an order, and clear the cart.
      operationId: checkoutCart
      parameters:
        - $ref: "#/components/parameters/cartParam"
      responses:
        "201":
          description: Successfully Checked Out Cart
          content:
            application/json:
              examples:
                cart:
                  value: |-
                    {
                        "message": "Order Submitted Successfully",
                        "order": {
                            "order_number": 25,
                            "total_price": "219.99",
                            "status": "Complete"
                        }
                    }
        "400":
          description: Missing Address Error
        "401":
          description: Authorization Error
        "500":
          description: Internal Server Error
      tags:
        - Carts
  /api/v1/orders:
    get:
      summary: Get All of the Orders' Data (ADMIN)
      description: Retrieve all of the users' orders information from the database.
      operationId: getOrders
      responses:
        "200":
          description: Successfully Retrieved Orders' Data
        "401":
          description: Authorization Error
        "500":
          description: Internal Server Error
      tags:
        - Orders
  /api/v1/orders/{user_id}:
    get:
      summary: Get a User's Order History (ADMIN / OWNER)
      description: Retrieve a user's order history matching the provided user ID.
      operationId: getOrdersByUserId
      parameters:
        - $ref: "#/components/parameters/userParam"
      responses:
        "200":
          description: Successfully Retrieved User's Order Data
          content:
            application/json:
              examples:
                orders:
                  value: |-
                    [
                        {
                            "order_number": 26,
                            "order_id": "a5df7c77-fef5-4c8b-b66f-ba542c61b1b3",
                            "total_price": "999.99",
                            "status": "Complete",
                            "created_at": "2022-04-23T05:34:12.497Z"
                        },
                        {
                            "order_number": 25,
                            "order_id": "e02e45e5-5635-4e16-b1bb-d93145373dc6",
                            "total_price": "219.99",
                            "status": "Complete",
                            "created_at": "2022-04-23T04:59:52.879Z"
                        }
                    ]
        "401":
          description: Authorization Error
        "404":
          description: Not Found Error
        "500":
          description: Internal Server Error
      tags:
        - Orders
  /api/v1/orders/details/{order_id}:
    get:
      summary: Get a User's Order Details (OWNER)
      description: Retrieve a user's order details matching the provided order ID.
      operationId: getOrderProducts
      parameters:
        - $ref: "#/components/parameters/orderParam"
      responses:
        "200":
          description: Successfully Retrieved Order Details
          content:
            application/json:
              examples:
                order:
                  value: |-
                    [
                        {
                            "product_id": "81294da5-f1fe-4d5d-8da2-27c8a4dd0a38",
                            "name": "Samsung Galaxy Note 8",
                            "quantity": 1,
                            "color": "Midnight Black",
                            "price": "219.99",
                            "img_urls": [
                                "https://res.cloudinary.com/de3kst1ud/image/upload/v1650202227/Galaxy%20Note%208/Samsung_Galaxy_Note_8_-_1_hbaavf.jpg",
                                "https://res.cloudinary.com/de3kst1ud/image/upload/v1650202229/Galaxy%20Note%208/Samsung_Galaxy_Note_8_-_2_vr91ku.jpg",
                                "https://res.cloudinary.com/de3kst1ud/image/upload/v1650202227/Galaxy%20Note%208/Samsung_Galaxy_Note_8_-_3_euojga.jpg"
                            ]
                        }
                    ]
        "401":
          description: Authorization Error
        "500":
          description: Internal Server Error
      tags:
        - Orders
  /api/v1/orders/review/{order_id}:
    get:
      summary: Get a User's Order Details as Admin (ADMIN)
      description: >-
        Allow Admin to retrieve a user's order details matching the provided
        order ID.
      operationId: getOrderProductsAsAdmin
      parameters:
        - $ref: "#/components/parameters/orderParam"
      responses:
        "200":
          description: Successfully Retrieved Order Details
          content:
            application/json:
              examples:
                order:
                  value: |-
                    [
                        {
                            "product_id": "81294da5-f1fe-4d5d-8da2-27c8a4dd0a38",
                            "name": "Samsung Galaxy Note 8",
                            "quantity": 1,
                            "color": "Midnight Black",
                            "price": "219.99",
                            "img_urls": [
                                "https://res.cloudinary.com/de3kst1ud/image/upload/v1650202227/Galaxy%20Note%208/Samsung_Galaxy_Note_8_-_1_hbaavf.jpg",
                                "https://res.cloudinary.com/de3kst1ud/image/upload/v1650202229/Galaxy%20Note%208/Samsung_Galaxy_Note_8_-_2_vr91ku.jpg",
                                "https://res.cloudinary.com/de3kst1ud/image/upload/v1650202227/Galaxy%20Note%208/Samsung_Galaxy_Note_8_-_3_euojga.jpg"
                            ]
                        },
                        {
                            "product_id": "244bd087-1e78-47f5-a3da-d783744d19a3",
                            "name": "Samsung Galaxy Note 5",
                            "quantity": 1,
                            "color": "Black Sapphire",
                            "price": "159.99",
                            "img_urls": [
                                "https://res.cloudinary.com/de3kst1ud/image/upload/v1650201496/Galaxy%20Note%205/Samsung%20Galaxy%20Note%205%20-%201.jpg",
                                "https://res.cloudinary.com/de3kst1ud/image/upload/v1650201732/Galaxy%20Note%205/Samsung_Galaxy_Note_5_-_2_hlknq6.jpg",
                                "https://res.cloudinary.com/de3kst1ud/image/upload/v1650201730/Galaxy%20Note%205/Samsung_Galaxy_Note_5_-_3_pwvy9t.jpg"
                            ]
                        }
                    ]
        "401":
          description: Authorization Error
        "404":
          description: Not Found Error
        "500":
          description: Internal Server Error
      tags:
        - Orders
components:
  schemas:
...
    Cart:
      type: object
      required:
        - product_id
        - quantity
        - color
      properties:
        product_id:
          type: string
          description: The ID of the product
          example: bd38abae-420a-4fbf-97f1-fbb316726e1f
        quantity:
          type: integer
          description: The quantity of the product
          example: 1
        color:
          type: string
          description: The color of the product
          example: Mystic Black
  parameters:
    userParam:
      name: user_id
      in: path
      description: ID of the user
      required: true
      schema:
        type: string
    productParam:
      name: product_id
      in: path
      description: ID of the product
      required: true
      schema:
        type: string
    cartParam:
      name: cart_id
      in: path
      description: ID of the cart
      required: true
      schema:
        type: string
    orderParam:
      name: order_id
      in: path
      description: ID of the order
      required: true
      schema:
        type: string